+++ src/main/java/br/com/fiap/postech/orders/interfaces/dto/OrderItemResponseDTO.java
@@ -1,12 +1,13 @@
 package br.com.fiap.postech.orders.interfaces.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 record OrderItemResponseDTO(
         UUID productId,
         int quantity,
-        double unitPrice,
-        double totalPrice
+        BigDecimal unitPrice,
+        BigDecimal totalPrice
 ) {
     public static OrderItemResponseDTO fromDomain(br.com.fiap.postech.orders.domain.entities.OrderItem item) {
         return new OrderItemResponseDTO(
